Update ogimet_daily.R
I experienced a few (random) errors and was frustrated that I could not recover data already download before the error occurred. So I modified the code accordingly by using 'next' or 'break'.

I also propose an improvement of the sequence of periods for a breakdown by month strictly (i.e. from 1 to 28, 30 or 31). In particular, an error is possible during a period straddling two years.